backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
google-genai
